LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Expected token '{'
S12.14_A16_T1

PASS Expected parsing failure

TEST COMPLETE
